api design changes for attachments
now attachment object is not used directly anymore, it is being constructed internally in `WithAttachment()` method
