+++ version
@@ -1,1 +1,1 @@
-0.5.2
+0.6.0
